[COS-1988] - phx presence for org channel
